backend/pyproject.toml
# This file was generated from bswck/skeleton@57cf553.
# Instead of changing this particular file, you might want to alter the template:
# https://github.com/bswck/skeleton/tree/57cf553/project/pyproject.toml.jinja

[tool.poetry]
name = "sharkservers-api"
version = "1.2.8"
description = "sharkservers-api"
authors = ["Qwizi <ciolek.adrian@protonmail.com>"]
license = "GPL-3.0"
readme = "README.md"
packages = [{ include = "sharkservers/" }]
homepage = "https://github.com/Qwizi/sharkservers-api"

[tool.poetry.urls]
Documentation = "https://sharkservers-api.readthedocs.io/en/latest/"
Issues = "https://github.com/Qwizi/sharkservers-api/issues"
Coverage = "https://coverage-badge.samuelcolvin.workers.dev/redirect/Qwizi/sharkservers-api"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
uvicorn = {extras = ["all"], version = "^0.25.0"}
fastapi-limiter = "^0.1.6"
ormar = "^0.12.2"
fastapi = "0.109.0"
steam = "^1.4.4"
fastapi-pagination = "^0.12.14"
fastapi-mail = "1.3.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
aiocron = "^1.8"
redis = "^5.0.1"
python-multipart = "^0.0.6"
httpx = "0.23.0"
sourcemod-api-client = "^0.0.1"
wsproto = "^1.2.0"
fastapi-events = "^0.10.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
fakeredis = "^2.20.1"
alembic = "^1.13.1"
python-dotenv = "^1.0.0"
pillow = "^10.2.0"
broadcaster = "^0.2.0"
asyncio-redis = "^0.16.0"
stripe = "^7.12.0"
resend = "^0.7.2"
fastapi-uuidbase62 = "^0.2"

[tool.poetry.group.dev-skeleton.dependencies]
# This dependency group was generated from bswck/skeleton@57cf553.
# Instead of changing this particular file, you might want to alter the template:
# https://github.com/bswck/skeleton/tree/57cf553/pyproject.toml
mypy = "^1.7.0"
ruff = "^0.1.11"
towncrier = "^23.11.0"
pytest = "^7.4.3"
coverage = "^7.3.2"
poethepoet = "^0.24.3"
pre-commit = "^3.5.0"

keyring = "^24.3.0"
[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
markdown-exec = ">=1.3.0"
mkdocs-material = ">=8.2,<10.0.0"
mkdocstrings = { version = ">=0.19.0", extras = ["python"] }


[tool.poetry.group.dev.dependencies]
pytest-dotenv = "^0.5.2"
pytest-env = "^1.1.3"
aiosqlite = "^0.19.0"
lupa = "^2.0"
trio = "^0.24.0"

[tool.poe.tasks]
test = "pytest -v"
lint = "ruff check ."
skeleton = "scripts/skeleton.57cf553.bash"
check = [
    { ref="test" },
    { ref="lint" },
]
release.script = "scripts.release:main"

[tool.poe.tasks.added]
shell = "towncrier create $TICKET.added.md --edit"
args = [{name = "TICKET", positional = true}]

[tool.poe.tasks.changed]
shell = "towncrier create $TICKET.changed.md --edit"
args = [{name = "TICKET", positional = true}]

[tool.poe.tasks.fixed]
shell = "towncrier create $TICKET.fixed.md --edit"
args = [{name = "TICKET", positional = true}]

[tool.poe.tasks.deprecated]
shell = "towncrier create $TICKET.deprecated.md --edit"
args = [{name = "TICKET", positional = true}]

[tool.poe.tasks.removed]
shell = "towncrier create $TICKET.removed.md --edit"
args = [{name = "TICKET", positional = true}]

[tool.poe.tasks.security]
shell = "towncrier create $TICKET.security.md --edit"
args = [{name = "TICKET", positional = true}]

[tool.ruff]
exclude = ["tests/", "scripts/", "migrations/", "sharkservers/apps/", "sharkservers/chat/", "sharkservers/subscryptions/", "sharkservers/services.py"]
# To discuss the presented rationales, contact the author (bswck).
select = ["ALL"]
ignore = [
    # Description: Missing type annotation for self in method
    # Rationale: It is mostly redundant, as is for ANN102.
    "ANN101",

    # Description: Missing type annotation for cls in method
    # Rationale: It is mostly redundant, as is for ANN101.
    "ANN102",

    # Description: Dynamically typed expressions (typing.Any) are disallowed
    # Rationale:
    #  We use Any to allow user to constrain return types of our functions on their own.
    #  For example, having a function `def foo() -> Any: ...` allows user to write
    #  `my_foo: int = foo()` and have it passed type checking, since `Any` disables
    #  type checking completely.
    "ANN401",

    # Description: 1 blank line required before class docstring
    # Rationale: Remove the warning -- D211 (no-blank-line-before-class) preferred.
    "D203",

    # Description: Multi-line docstring summary should start at the second line
    # Rationale: Remove the warning -- D213 (multi-line-summary-second-line) preferred.
    "D212",

    # Description: Line contains TODO, consider resolving the issue
    # Rationale: Not appropriate for the project.
    "FIX002",
    "TCH001",
    "TCH002",
    "E501",
    "EXE002",
    "B008",
    "A003",
]
[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = ["pydantic.BaseModel", "sqlalchemy.orm.DeclarativeBase"]

[tool.mypy]
strict = true

[tool.isort]
profile = "black"

[tool.towncrier]
directory = "changes"
package = "sharkservers"
filename = "CHANGELOG.md"
start_string = "<!-- insertion marker -->\n"
underlines = ["", "", ""]  # We use Markdown
title_format = "## [{version}](https://github.com/Qwizi/sharkservers-api/tree/{version}) ({project_date})"
issue_format = "[#{issue}](https://github.com/Qwizi/sharkservers-api/issues/{issue})"

[[tool.towncrier.type]]
directory = "security"
name = "Security"
showcontent = true

[[tool.towncrier.type]]
directory = "removed"
name = "Removed"
showcontent = true

[[tool.towncrier.type]]
directory = "deprecated"
name = "Deprecated"
showcontent = true

[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Fixed"
showcontent = true

[tool.coverage.report]
omit = ["tests/*"]
# fail_under = 96
exclude_also = [
    # Be sure to include lines that are not meant to be covered
    # by tests or encountered by the users at runtime.
    "raise NotImplementedError",          # Do not cover abstract methods.
    "if ((t|typing)\\.?)?TYPE_CHECKING:", # Do not cover type checking imports.
    "@((t|typing)\\.?)?overload",         # Do not cover overload definitions.
    "\\(((t|typing)\\.?)?Protocol\\):$",  # Do not cover entire protocol definitions.
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
